navigation between activities, event listeners with images refactoring classes and layouts.
