+++ app/src/main/java/com/example/eletriccarapp/presentation/CalculateAutonomyActivity.kt
@@ -1,36 +1,35 @@
-package com.example.eletriccarapp
+package com.example.eletriccarapp.presentation
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
 import android.widget.EditText
+import android.widget.ImageView
 import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
+import com.example.eletriccarapp.R
 
-class MainActivity : AppCompatActivity() {
+class CalculateAutonomyActivity : AppCompatActivity() {
 
     lateinit var price: EditText
-    lateinit var btnCalculate: Button
     lateinit var kmTraveled: EditText
     lateinit var result: TextView
+    lateinit var btnCalculate: Button
+    lateinit var btnClose: ImageView
+
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.activity_calculate_autonomy)
         setupView()
-        setupListeners()
+        setupListener()
     }
 
     private fun setupView() {
         price = findViewById(R.id.et_preco_kwh)
-        btnCalculate = findViewById(R.id.btn_calcular)
         kmTraveled = findViewById(R.id.et_km_percorrido)
         result = findViewById(R.id.tv_resultado)
-
-    }
-
-    private fun setupListeners() {
-        btnCalculate.setOnClickListener {
-            calculate()
-        }
+        btnCalculate = findViewById(R.id.btn_calcular)
+        btnClose = findViewById(R.id.iv_close)
     }
 
     private fun calculate() {
@@ -38,6 +37,17 @@ class MainActivity : AppCompatActivity() {
         val km = kmTraveled.text.toString().toFloat()
         val resultPrice = price / km
         result.text = resultPrice.toString()
+    }
+
+    private fun setupListener(){
+        btnCalculate.setOnClickListener {
+            calculate()
+        }
+
+        btnClose.setOnClickListener {
+            // Desempilha e volta para tela anterior
+            finish()
+        }
     }
