+++ app/src/main/res/values/strings.xml
@@ -10,4 +10,6 @@
     <string name="preco_kwh">Preço Kwh</string>
     <string name="km_percorrido">Km Percorrido</string>
     <string name="calcular">Calcular</string>
+    <string name="ir_para_calculadora">Ir para calculadora</string>
+    <string name="calcular_autonomia">Calcular Autonomia</string>
 </resources>
